apps/worker/sandbox-ai/requirements.txt
onnxruntime==1.22.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3
